refactor(snmptraps): migrate oidresolver component to v2 structure - Move interface and types to oidresolver/def/ - Move implementation to oidresolver/impl/ with Requires/Provides/NewComponent - Add oidresolver/fx/ with fxutil.ProvideComponentConstructor - Update import paths in formatter and server
