+++ comp/snmptraps/oidresolver/def/component.go
