+++ comp/snmptraps/oidresolver/def/traps_db.go
