+++ comp/snmptraps/server/serverimpl/server.go
@@ -24,7 +24,7 @@ import (
 	"github.com/DataDog/datadog-agent/comp/snmptraps/forwarder/forwarderimpl"
 	"github.com/DataDog/datadog-agent/comp/snmptraps/listener"
 	"github.com/DataDog/datadog-agent/comp/snmptraps/listener/listenerimpl"
-	"github.com/DataDog/datadog-agent/comp/snmptraps/oidresolver/oidresolverimpl"
+	oidresolverfx "github.com/DataDog/datadog-agent/comp/snmptraps/oidresolver/fx"
 	"github.com/DataDog/datadog-agent/comp/snmptraps/server"
 	"github.com/DataDog/datadog-agent/comp/snmptraps/status/def"
 	statusimpl "github.com/DataDog/datadog-agent/comp/snmptraps/status/impl"
@@ -108,7 +108,7 @@ func newServer(lc fx.Lifecycle, deps dependencies) provides {
 		formatterimpl.Module(),
 		forwarderimpl.Module(),
 		listenerimpl.Module(),
-		oidresolverimpl.Module(),
+		oidresolverfx.Module(),
 		fx.Invoke(func(_ forwarder.Component, _ listener.Component) {}),
 	)
 	server := &TrapsServer{app: app, stat: stat}
